+++ modules/segmentshortestpaths/core/segmtopologicalshortestpath.cpp
@@ -29,13 +29,13 @@ AnalysisResult SegmentTopologicalShortestPath::run(Communicator *comm) {
     }
 
     AttributeTable &attributes = m_map.getAttributeTable();
-    int shapeCount = m_map.getShapeCount();
+    size_t shapeCount = m_map.getShapeCount();
 
     std::string colText = "Topological Shortest Path Depth";
-    int depth_col = attributes.insertOrResetColumn(colText);
+    size_t depth_col = attributes.insertOrResetColumn(colText);
     result.addAttribute(colText);
     colText = "Topological Shortest Path Order";
-    int path_col = attributes.insertOrResetColumn(colText);
+    size_t path_col = attributes.insertOrResetColumn(colText);
     result.addAttribute(colText);
 
     // record axial line refs for topological analysis
